Revert "Merged CommentsList and TeamComments"
This reverts commit 7531ff24c4ad27402193c86575e608b1b6413055.
