+++ frontend/src/core/comments/index.js
@@ -2,3 +2,4 @@
 
 export { default as AddComment } from './components/AddComment';
 export { default as Comment } from './components/Comment';
+export { default as CommentsList } from './components/CommentsList';
